+++ arkimapslib/steps.py
@@ -3,6 +3,7 @@ from typing import Dict, Any, Optional, Set, Tuple
 
 # if TYPE_CHECKING:
 from . import inputs
+from .lint import Lint
 # Used for kwargs-style dicts
 Kwargs = Dict[str, Any]
 
@@ -41,6 +42,20 @@ class Step:
         self.params = self.compile_args(step_config, params if params is not None else {})
         if bool(self.params.get("skip")):
             raise StepSkipped()
+
+    @classmethod
+    def lint(
+            cls,
+            lint: Lint, *,
+            defined_in: str,
+            name: str,
+            step: str,
+            **kwargs):
+        """
+        Consistency check the given input arguments
+        """
+        for k, v in kwargs.items():
+            lint.warn_recipe_step(f"Unknown parameter: {k!r}", defined_in=defined_in, name=name, step=step)
 
     def as_magics_macro(self) -> Tuple[str, Dict[str, Any]]:
         """
@@ -103,6 +118,14 @@ class MagicsMacro(Step):
     """
     macro_name: str
 
+    @classmethod
+    def lint(
+            cls,
+            lint: Lint, *,
+            params: Optional[Dict[str, Any]] = None,
+            **kwargs):
+        super().lint(lint, **kwargs)
+
     def as_magics_macro(self) -> Tuple[str, Dict[str, Any]]:
         return self.macro_name, self.params.get("params", {})
 
@@ -154,6 +177,30 @@ class AddContour(MagicsMacro):
         },
     }
 
+    @classmethod
+    def lint(
+            cls,
+            lint: Lint, *,
+            params: Optional[Dict[str, Any]] = None,
+            **kwargs):
+        super().lint(lint, params=params, **kwargs)
+        cll = params.get("contour_level_list")
+        cscl = params.get("contour_shade_colour_list")
+        if cll is not None and cscl is not None:
+            ok = True
+            if not isinstance(cll, list):
+                lint.warn_recipe_step(f"contour_level_list {cll!r} is not a list", **kwargs)
+                ok = False
+            if not isinstance(cscl, list):
+                lint.warn_recipe_step(f"contour_level_list {cll!r} is not a list", **kwargs)
+                ok = False
+            if ok:
+                if len(cll) != len(cscl) + 1:
+                    lint.warn_recipe_step(
+                        f"contour_level_list has {len(cll)} items while"
+                        f" contour_shade_colour_list has {len(cscl)} items",
+                        **kwargs)
+
 
 class AddWind(MagicsMacro):
     """
@@ -174,10 +221,11 @@ class AddGrid(MagicsMacro):
     }
 
 
-class AddCoastlinesFg(Step):
+class AddCoastlinesFg(MagicsMacro):
     """
     Add foreground coastlines
     """
+    macro_name = "mcoast"
     defaults = {
         "params": {
             "map_coastline_sea_shade_colour": "#f2f2f2",
@@ -189,15 +237,12 @@ class AddCoastlinesFg(Step):
         },
     }
 
-    def as_magics_macro(self) -> Tuple[str, Dict[str, Any]]:
-        # TODO: if implementation stays like this, make this a subclass of MagicsMacro
-        return "mcoast", self.params.get("params", {})
 
-
-class AddBoundaries(Step):
+class AddBoundaries(MagicsMacro):
     """
     Add political boundaries
     """
+    macro_name = "mcoast"
     defaults = {
         "params": {
             'map_boundaries': "on",
@@ -208,10 +253,6 @@ class AddBoundaries(Step):
             'map_administrative_boundaries': "on",
         },
     }
-
-    def as_magics_macro(self) -> Tuple[str, Dict[str, Any]]:
-        # TODO: if implementation stays like this, make this a subclass of MagicsMacro
-        return "mcoast", self.params.get("params", {})
 
 
 class AddGrib(Step):
@@ -236,6 +277,15 @@ class AddGrib(Step):
                 self.params["params"] = params = {}
             for k, v in self.grib_input.info.mgrib.items():
                 params.setdefault(k, v)
+
+    @classmethod
+    def lint(
+            cls,
+            lint: Lint, *,
+            grib: str,
+            params: Optional[Dict[str, Any]] = None,
+            **kwargs):
+        super().lint(lint, **kwargs)
 
     def as_magics_macro(self) -> Tuple[str, Dict[str, Any]]:
         params = dict(self.params.get("mgrib", {}))